requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.3.2
pydantic==2.6.1
pydantic-settings==2.2.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-slugify==8.0.4
httpx==0.26.0
markdown2==2.4.13
bleach==6.1.0
